- ベンチマークの実行時間を一定にする改良 	- ベンチマーク、探索部の改良により、実行時間が変動するのはあまり良くない性質 	- 固定時間で計測してnpsで比較すべき。 	- 現状、デフォルトでは固定depthになっているが、これを固定timeに変更する。 	- cf. https://github.com/yaneurao/YaneuraOu/issues/195 	- ベンチマーク局面、一つ追加。4局面×15秒 = 60秒。
- learnerでコンパイル時にwarning出てたの修正。

+++ source/learn/learner.cpp
@@ -1354,7 +1354,7 @@ struct SfenReader
 
 			PSVector sfens(SFEN_READ_SIZE);
 			// 次にこの位置から読み込む。
-			int sfens_read_offset = 0;
+			size_t sfens_read_offset = 0;
 
 			// ファイルバッファにファイルから読み込む。
 			while (sfens_read_offset < SFEN_READ_SIZE)
